Refactor types into separate file
